Move src/components/inbox/ to src/apps/mydb/inbox/
