+++ app/packs/src/apps/mydb/inbox/AttachmentContainer.js
@@ -7,7 +7,7 @@ import InboxActions from 'src/stores/alt/actions/InboxActions';
 import DragDropItemTypes from 'src/components/DragDropItemTypes';
 import Utils from 'src/utilities/Functions';
 
-import MoveToAnalysisButton from 'src/components/inbox/MoveToAnalysisButton';
+import MoveToAnalysisButton from 'src/apps/mydb/inbox/MoveToAnalysisButton';
 import InboxStore from 'src/stores/alt/stores/InboxStore';
 import ArrayUtils from 'src/utilities/ArrayUtils';
